+++ packages/utils/src/index.ts
@@ -20,3 +20,4 @@ export * from "./asset";
 export * from "./chainData";
 export * from "./time";
 export * from "./gasEstimates";
+export * from "./fallbackSubgraph";